requirements.txt
julep
fastapi
uvicorn
python-dotenv 
pydantic
PyYAML